Move Dart extensions to new `extensions/` folder
