+++ lib/core/utils/extensions/theme_mode_ext.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart' show IconData, Icons, ThemeMode;
 import 'package:flutter/widgets.dart';
-import 'package:vidya_music/core/utils/build_context_l10n_ext.dart';
+import 'package:vidya_music/core/utils/extensions/build_context_l10n_ext.dart';
 
 extension ThemeModeExt on ThemeMode {
   String l10n(BuildContext context) {
